pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "python-dabmux"
version = "0.1.0"
description = "DAB/DAB+ multiplexer in Python"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "python-dabmux contributors"}
]
dependencies = [
    "structlog>=23.0",
    "pyyaml>=6.0",
    "pyzmq>=25.0.0",  # Priority 6: Remote control
    "watchdog>=3.0.0",  # MOT carousel hot-reload
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocs-mermaid2-plugin>=1.1",
    "pymdown-extensions>=10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=dabmux --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
